requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.4
django-material-admin==1.8.6
djangorestframework==3.14.0
gunicorn==20.1.0

Pillow==9.3.0
psycopg2==2.9.5

sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
